requirements.txt
asgiref==3.5.0
crispy-bootstrap5==0.6
dj-database-url==0.5.0
Django==3.1.14
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
environs==8.0.0
gunicorn==19.9.0
marshmallow==3.14.1
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.1.0